#5833 Separated python package installations by adding a new variable. Also separated shell scripts into different layers in order to utilize caching when building locally and save time during failed builds. Also updated script for openjdk to install openjdk 8 for CentOS 6 since it doesnt have OpenJDK 11 available. Python separation should get some additional fixes to also avoid the python bundles to be installed but that will done in at a later time.

+++ BaseDockerImages/trunk/scripts/setup_jcc.sh
@@ -14,29 +14,31 @@
 
 set -e
 
-# find Java location containing jni.h, variable JCC_JDK needs to be set to install jcc
-# also grep for the installed java jdk package
-if [ -f /etc/centos-release ] || [ -f /etc/fedora-release ];
-then
-    export JCC_JDK=$(find /usr -type f -name "jni.h" | cut -d '/' -f-5 | grep "1.8.0")
-else
-    export JCC_JDK=$(find /usr -type f -name "jni.h" | cut -d '/' -f-5 | grep "java-8")
+if [ ! -z "${PYTHON_ENABLED}" ]; then
+    # find Java location containing jni.h, variable JCC_JDK needs to be set to install jcc
+    # also grep for the installed java jdk package
+    if [ -f /etc/centos-release ] || [ -f /etc/fedora-release ];
+    then
+        export JCC_JDK=$(readlink -f /usr/bin/java | cut -d '/' -f-5)
+    else
+        export JCC_JDK=$(find /usr -type f -name "jni.h" | cut -d '/' -f-5 | grep "java-8")
+    fi
+
+    echo "GOT JCC_JDK="$JCC_JDK
+    lines=$(echo $JCC_JDK | wc -l)
+
+    if [ $lines -eq 0 ];
+    then
+        echo -e "\e[31m" "setup_jcc: Could not find jni.h" "\e[0m"
+        exit 1
+    fi
+
+    pip install jcc==2.23
+
+    JCC_INSTALL_DIR=$(find /usr -type d -name jcc | sed -n 1p)
+    cp ${DOCKER_SRC_DIR}/jcc.patch ${JCC_INSTALL_DIR}
+    cd ${JCC_INSTALL_DIR}
+    patch < jcc.patch
+
+    rm -i jcc.patch
 fi
-
-echo "GOT JCC_JDK="$JCC_JDK
-lines=$(echo $JCC_JDK | wc -l)
-
-if [ $lines -eq 0 ];
-then
-    echo -e "\e[31m" "setup_jcc: Could not find jni.h" "\e[0m"
-    exit 1
-fi
-
-pip install jcc==2.23
-
-JCC_INSTALL_DIR=$(find /usr -type d -name jcc | sed -n 1p)
-cp ${DOCKER_SRC_DIR}/jcc.patch ${JCC_INSTALL_DIR}
-cd ${JCC_INSTALL_DIR}
-patch < jcc.patch
-
-rm -i jcc.patch
